Move test to isolate it
